isolated test in seperate suite to avoid side effects
